miscellaneous portability improvements
git-svn-id: http://arthurdejong.org/svn/nss-pam-ldapd/nss-ldapd@742 ef36b2f9-881f-0410-afb5-c4e39611909c

+++ nslcd/group.c
@@ -99,7 +99,7 @@ static int mkfilter_group_bygid(gid_t gid,
   return mysnprintf(buffer,buflen,
                     "(&%s(%s=%d))",
                     group_filter,
-                    attmap_group_gidNumber,gid);
+                    attmap_group_gidNumber,(int)gid);
 }
 
 /* create a search filter for searching a group entry
